Changed the sample script again to set up more dynamic scripts.
I decided that the previous method, which uses primarily HTML tags, would be too limiting when animated sequences would come into play. So I took a style inspired by SQL, which the major difference being that it is specialized for video creation, not database management.

Key words are subject to change, and functionality for the animation will come after this sample script succeeds. However, this will more clearly set up that functionality when it is implemented.